add global depends for settings; add todo
