+++ fastapi/api/v1/login.py
@@ -6,17 +6,13 @@ from pydantic import BaseModel
 from fastapi import APIRouter, Depends, HTTPException
 from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
 
+from ..config import Settings
 from ..model import User as UserInDB
 from ..util.hash import build_json_hash
 from ..util.jwt import TokenFactory
 
-SECRET_KEY = os.environ['SECRET_KEY']
-# TODO: read from app settings
-
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl='./api/v1/login')
-token_factory = TokenFactory(secret_key=os.environ['SECRET_KEY'])
-
 router = APIRouter()
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl='./api/v1/login')
 
 
 class User(BaseModel):
@@ -29,6 +25,10 @@ class User(BaseModel):
         orm_mode = True
 
 
+async def get_token_factory(settings: Settings = Depends()) -> TokenFactory:
+    return TokenFactory(secret_key=settings.secret_key)
+
+
 def verify_password(plain_password: str, hashed_password: str, salt: str) -> bool:
     return build_json_hash(data={'password': plain_password}, salt=salt) == hashed_password
 
@@ -39,7 +39,7 @@ def authenticate_user(username: str, password: str) -> Optional[UserInDB]:
         return user
 
 
-async def get_current_user(token: str = Depends(oauth2_scheme)) -> UserInDB:
+async def get_current_user(token: str = Depends(oauth2_scheme), token_factory: TokenFactory = Depends(get_token_factory)) -> UserInDB:
     to_raise = HTTPException(status_code=401, detail='could not validate token')
     token_data = await token_factory.parse(token)
     if token_data:
@@ -60,7 +60,7 @@ async def get_current_active_user(current_user: User = Depends(get_current_user)
 
 
 @router.post('/login')
-async def login(form_data: OAuth2PasswordRequestForm = Depends()):
+async def login(form_data: OAuth2PasswordRequestForm = Depends(), token_factory: TokenFactory = Depends(get_token_factory)):
     user = authenticate_user(username=form_data.username, password=form_data.password)
     if not user:
         raise HTTPException(status_code=401, detail='incorrect username or password')
