+++ fastapi/main.py
@@ -1,9 +1,12 @@
 import uvicorn
-from config import Settings
-from factory import create_app
+
+from api.config import Settings
+from api.factory import create_app
 
 settings = Settings()
 app = create_app(settings)
 
 if __name__ == '__main__':
     uvicorn.run('main:app', host=settings.service_host, port=settings.service_port, reload=settings.debug)
+
+# TODO: revisit project structure
